requirements.txt
flask
python-dotenv
boto3 